Buffer now uses 32-bit int to represent size. updated map_value_to_str() and MapKey::argument_value() to use the struct information for navigating to data

+++ src/ast/codegen_llvm.cpp
@@ -567,7 +567,7 @@ void CodegenLLVM::visit(Call &call)
       auto &arg = *call.vargs->at(0);
       fixed_buffer_length = arg.type.GetNumElements() *
                             arg.type.GetElementTy()->size;
-      length = b_.getInt64(fixed_buffer_length);
+      length = b_.getInt32(fixed_buffer_length);
     }
 
     auto elements = AsyncEvent::Buf().asLLVMType(b_, fixed_buffer_length);
